CR+LF functionality in Terminal
Terminal now handles carriage return/linefeed in text files sent from
the server. Temporary payload <P> element is still included in the
terminal along with the timing code. This will be disabled/removed in
the next commit.

Minor changes to commands.txt. This will be fleshed out into a proper
json struct with various attributes in a future rev.